build-tsc build script cleanup
With this change, both `npm run build` and `npm run build-ts` put their output in the __`lib`__ directory.  This allows us to compare behavior between the two versions of the library using `npm test` in either case.

Right now, I'm getting 0 failures with the `build` (babel) version of the library, and 20 failures with the `build-tsc` (TypeScript) version.

- clean out the built directory before tsc
- remove hack to ignore erros from tsc
- checkin vscode launch/debug files
- compile _only_ the `src` directory with tsc
- switch back to default import of sinonChai in test.spec.js