Support Yarn
Due to a complicated bug, npm is acting strange for me.   I've found the Yarn package manager doesn't have the same prolem.   Added a yarn.lock file, and a note in README.md.

Having both a yarn and npm lock file seems like it might be undesirable long term, but we can reexamine this once some one of the npm/VSCode/typescript bugs are resolved.